+++ src/Dto/CheckSummary.php
@@ -2,6 +2,8 @@
 
 namespace OhDear\PhpSdk\Dto;
 
+use OhDear\PhpSdk\Enums\CheckResult;
+
 class CheckSummary
 {
     public function __construct(
@@ -9,11 +11,16 @@ class CheckSummary
         public ?string $summary,
     ) {}
 
+    public function checkResult(): ?CheckResult
+    {
+        return CheckResult::tryFrom($this->result);
+    }
+
     public static function fromResponse(array $data): self
     {
         return new self(
             result: $data['result'],
-            summary: $data['summary'],
+            summary: $data['summary'] ?? null,
         );
     }
 }